Add an awkward Dhall.Parser.Token module
