+++ src/Dhall/Parser/Token.hs
@@ -1,0 +1,140 @@
+{-# LANGUAGE OverloadedStrings #-}
+-- | Parse Dhall tokens. Even though we don't have a tokenizer per-se this
+---  module is useful for keeping some small parsing utilities.
+module Dhall.Parser.Token where
+
+import           Dhall.Parser.Combinators
+
+import Control.Applicative (Alternative(..), liftA2, optional)
+import Control.Exception (Exception)
+import Control.Monad (MonadPlus)
+import Data.ByteArray.Encoding (Base(..))
+import Data.Data (Data)
+import Data.Functor (void)
+import Data.HashMap.Strict.InsOrd (InsOrdHashMap)
+import Data.Scientific (Scientific)
+import Data.Semigroup (Semigroup(..))
+import Data.Sequence (ViewL(..))
+import Data.Set (Set)
+import Data.String (IsString(..))
+import Data.Text (Text)
+import Data.Void (Void)
+import Dhall.Core
+import Formatting.Buildable (Buildable(..))
+import Numeric.Natural (Natural)
+import Prelude hiding (const, pi)
+import Text.Parser.Combinators (choice, try, (<?>))
+import Text.Parser.Token (TokenParsing(..))
+
+import qualified Control.Monad
+import qualified Crypto.Hash
+import qualified Data.ByteArray.Encoding
+import qualified Data.ByteString
+import qualified Data.Char
+import qualified Data.HashMap.Strict.InsOrd
+import qualified Data.HashSet
+import qualified Data.List
+import qualified Data.List.NonEmpty
+import qualified Data.Sequence
+import qualified Data.Set
+import qualified Data.Text
+import qualified Data.Text.Encoding
+import qualified Text.Megaparsec
+import qualified Text.Megaparsec.Char
+import qualified Text.Parser.Char
+import qualified Text.Parser.Combinators
+import qualified Text.Parser.Token
+import qualified Text.Parser.Token.Style
+
+
+whitespace :: Parser ()
+whitespace = Text.Parser.Combinators.skipMany whitespaceChunk
+
+alpha :: Char -> Bool
+alpha c = ('\x41' <= c && c <= '\x5A') || ('\x61' <= c && c <= '\x7A')
+
+digit :: Char -> Bool
+digit c = '\x30' <= c && c <= '\x39'
+
+hexdig :: Char -> Bool
+hexdig c =
+        ('0' <= c && c <= '9')
+    ||  ('A' <= c && c <= 'F')
+    ||  ('a' <= c && c <= 'f')
+
+whitespaceChunk :: Parser ()
+whitespaceChunk =
+    choice
+        [ void (Text.Parser.Char.satisfy predicate)
+        , void (Text.Parser.Char.text "\r\n")
+        , lineComment
+        , blockComment
+        ] <?> "whitespace"
+  where
+    predicate c = c == ' ' || c == '\t' || c == '\n'
+
+hexNumber :: Parser Int
+hexNumber = choice [ hexDigit, hexUpper, hexLower ]
+  where
+    hexDigit = do
+        c <- Text.Parser.Char.satisfy predicate
+        return (Data.Char.ord c - Data.Char.ord '0')
+      where
+        predicate c = '0' <= c && c <= '9'
+
+    hexUpper = do
+        c <- Text.Parser.Char.satisfy predicate
+        return (10 + Data.Char.ord c - Data.Char.ord 'A')
+      where
+        predicate c = 'A' <= c && c <= 'F'
+
+    hexLower = do
+        c <- Text.Parser.Char.satisfy predicate
+        return (10 + Data.Char.ord c - Data.Char.ord 'a')
+      where
+        predicate c = 'a' <= c && c <= 'f'
+
+lineComment :: Parser ()
+lineComment = do
+    _ <- Text.Parser.Char.text "--"
+    Text.Parser.Combinators.skipMany notEndOfLine
+    endOfLine
+    return ()
+  where
+    endOfLine =
+            void (Text.Parser.Char.char '\n'  )
+        <|> void (Text.Parser.Char.text "\r\n")
+
+    notEndOfLine = void (Text.Parser.Char.satisfy predicate)
+      where
+        predicate c = ('\x20' <= c && c <= '\x10FFFF') || c == '\t'
+
+blockComment :: Parser ()
+blockComment = do
+    _ <- Text.Parser.Char.text "{-"
+    blockCommentContinue
+
+blockCommentChunk :: Parser ()
+blockCommentChunk =
+    choice
+        [ blockComment  -- Nested block comment
+        , character
+        , endOfLine
+        ]
+  where
+    character = void (Text.Parser.Char.satisfy predicate)
+      where
+        predicate c = '\x20' <= c && c <= '\x10FFFF' || c == '\n' || c == '\t'
+
+    endOfLine = void (Text.Parser.Char.text "\r\n")
+
+blockCommentContinue :: Parser ()
+blockCommentContinue = endOfComment <|> continue
+  where
+    endOfComment = void (Text.Parser.Char.text "-}")
+
+    continue = do
+        blockCommentChunk
+        blockCommentContinue
+
+
